wlcore: Add support for DT platform data
When running with DT, we no longer have a board file that can set up the
platform data for wlcore. Allow this data to be passed from DT.

Since some platforms use a gpio-irq, add support for passing either the
irq number or the gpio number. For the latter case, the driver will
request the gpio and convert it to the irq number. If an irq is
specified, it'll be used as is.

[Arik - the pdev_data pointer does not belong to us and is freed when
the device is released. Dereference to our private data first.]

[Eliad - Older kernels don't have the gpio field defined.
In order to be able to work against them, move the gpio
field to the end of the struct, and add a comment
to avoid using it if unsure.

(it is safe to use it if the whole struct is defined
by the driver, specifically in the case of device-tree
without platform_data defined in the kernel)]

Signed-off-by: Ido Yariv <ido@wizery.com>
Signed-off-by: Arik Nemtsov <arik@wizery.com>
Signed-off-by: Eliad Peller <eliad@wizery.com>