clk: add flags to distinguish xtal clocks
Add a flag that indicate whether the clock is a crystal or not.  Since
no clocks set this flag right now, include an additional flag that
indicates whether the type is set or not.  If the CLK_IS_TYPE_DEFINED
flag is not set, the value of the CLK_IS_TYPE_XTAL flag is undefined.
This ensures backwards compatibility.

Additionally, parse a new device tree binding in clk-fixed-rate to set
this flag.

Signed-off-by: Luciano Coelho <coelho <at> ti.com>

downloaded from https://lkml.org/lkml/2013/7/29/374 and adjusted to current
kernel

Signed-off-by: Radek Dostal <radek.dostal@streamunlimited.com>