arm: dts: dra7: Add syscon for smasw in VIP nodes
When using falling edge pixel clock with VIP, pclk needs to be inverted
inside SoC before it is given to the VIP block.

This is achieved by setting control module register "CTRL_CORE_SMA_SW1"
Add a syscon property for each VIP node to allow this to be set by the
driver.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>