ARM: i.MX6UL: Add cl-som-imx6ul platform driver.
Add cl-som-imx6ul platform driver.
This patch sets a correct value of the ENET1 reference
clock mode selection field (GPR1.13 = 1).
It is required in order to make FEC and PHY work together
with respect to the CompuLab board design.
In cl-som-imx6ul boards, ENET ref clock source is a kzs8041 phy.
Details in the IMX6ULRM.

Signed-off-by: Valentin Raevsky <valentin@compulab.co.il>