sil164: add dvi transmitter simple driver.
Add dvi transmitter simple driver.
cl-sb-som-imx6ul makes use of a sil164 dvi transmitter.
The transmitter works in the i2c mode.
The purpose of the driver is to turn on/off the transmitter and
set the correct init values to the controls' registers.

Signed-off-by: Valentin Raevsky <valentin@compulab.co.il>
Signed-off-by: Stephan Linz <linz@li-pro.net>

Cherry-picked: e0e40520c1d97ab0491238787d80745326a73672